refactor: add Portfolio model and update AppUser and Stock models
This commit adds the Portfolio model to the application. The AppUser and Stock models have been updated to include a list of Portfolios. This change allows users to have multiple portfolios and associate stocks with each portfolio.

Refactor the code to improve the organization and structure of the models.

Commands
dotnet ef migrations add PortfolioManyToMany
dotnet ef database update